# auto vectorize `std::count`