@AlexGuteniev 2024-05-17
The main difference between `iota` auto vectorization and this one is that `iota` auto vectorization does not attempt to handle large ranges which would wrap counter. This makes `iota` auto vectorization simpler.

I don't think `iota` should handle large ranges, because it is squirely to wrap counter (and UB for signed counter). Also it is not very efficient to use large `iota` arrays instead of `iota_view`.

In contrast, counting large number of small elements is equally reasonable as counting small numbers of them.